+++ server/db/db.js
@@ -13,9 +13,9 @@ if(process.env.LOGGING === 'true'){
   delete config.logging
 }
 
-// const db = new Sequelize(process.env.DATABASE_URL || 'postgres://postgres:JerryPine@localhost/basketball');
-const db = new Sequelize(
-	process.env.DATABASE_URL || "postgres://localhost/basketball", config
-);
+const db = new Sequelize(process.env.DATABASE_URL || 'postgres://postgres:JerryPine@localhost/basketball');
+// const db = new Sequelize(
+// 	process.env.DATABASE_URL || "postgres://localhost/basketball", config
+// );
 
 module.exports = db;
