+++ server/db/index.js
@@ -5,39 +5,10 @@ const db = require('./db');
 const Game = require('./models/Game');
 const Request = require('./models/Request');
 const UserGame = require('./models/UserGame');
+const User = require('./models/User');
 
 // Auth
 const bcrypt = require("bcrypt");
-
-//So we can include a ton more personal information but just wanted to get
-// us started
-const User = db.define(
-	"user",
-	{
-		email: {
-			type: STRING,
-		},
-		password: {
-			type: STRING,
-		},
-		name: {
-			type: STRING,
-		},
-		age: {
-			type: INTEGER,
-		},
-		height: {
-			type: STRING,
-		},
-		description: {
-			type: STRING,
-		},
-		photo: {
-			type: STRING,
-		},
-	},
-	{ timestamps: false }
-);
 
 // Salt passwords
 User.beforeCreate(async (user, options) => {
